confroid utils
la fonction qui utilise reflect pour avoir les champs + valeurs d'un objet et la fonction qui convertit un object reflected en objet passable dans un Bundle sont dans ConfroidUtils désormais.
Début de l'implémentation de ConfroidUtils avec les fonctions d'envoi d'une Config, de récupération d'une Config et de récupération de toutes les versions d'une Config (sauf qu'il n'y a pour l'instant aucun échange de requêtes, tokens etc..)
